+++ builtin/app/go/factory.go
@@ -1,0 +1,9 @@
+package goapp
+
+import (
+	"github.com/hashicorp/otto/app"
+)
+
+func AppFactory() app.App {
+	return &App{}
+}