+++ helper/vagrant/ssh.go
@@ -2,6 +2,7 @@ package vagrant
 
 import (
 	"io"
+	"log"
 	"os"
 	"os/exec"
 	"strings"
@@ -27,7 +28,8 @@ type SSHCache struct {
 func (c *SSHCache) Exec(cacheOkay bool) error {
 	// If we have the cache file, use that
 	if _, err := os.Stat(c.Path); err == nil {
-		cmd := exec.Command("ssh", "-F", c.Path, "default")
+		log.Printf("[DEBUG] ssh command: ssh -F " + c.Path + " default")
+		cmd := exec.Command("ssh", "-t", "-t", "-F", c.Path, "default")
 		cmd.Stdin = os.Stdin
 		cmd.Stdout = os.Stdout
 		cmd.Stderr = os.Stderr