+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 ## Next Version
 
+FEATURES:
 
+  * **Layered Dev Environments**: Dev environments are now layered. Each
+    layer is cached. When bringing up a new development environment, cached
+    layers are used to speed it up immensely.
+  * **App Type Plugins**: You can now add custom app types (or even override
+    Otto's built-in types) using app type plugins. Want to support a new
+    language? A new framework? App type plugins are for you.
+
+IMPROVEMENTS:
+
+  * app/ruby: Automatically detect desired Ruby version and install it [GH-293]
+  * app/ruby: Install PhantomJS when poltergeist is detected [GH-313]
+
+BUG FIXES:
+
+  * command/compile: compilation works if Appfile is a directory (it
+      ignores the directory and detects an Appfile) [GH-280]
 
 ## 0.1.2 (October 20, 2015)